requirements.txt
pycurl
pyyaml
git+git://github.com/dmall/jsonrpclib-py3.git